@@ -21,4 +21,4 @@ const users = [
   },
 ]
 
-module.exports = { users }
+export default users
